move testing to alternative dir, restore original struct
